+++ CHANGELOG.md
@@ -10,13 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- support for webexec clipboard commands, e.g. use `webexec copy < file.txt` to copy a file to the client's clipboard
+- Support for webexec clipboard commands, e.g. use `webexec copy < file.txt` to copy a file to the client's clipboard
 
 ## [1.9.2] - 2023/3/18
 
 ### Added
 
-- support command to send the logs to the server
+- Support command to send the logs to the server
 - 9000 more lines of scrollback and a new setting to control it
 
 ### Fixed