+++ CHANGELOG.md
@@ -15,13 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support for webexec clipboard commands, e.g. use `webexec copy < file.txt` to copy a file to the client's clipboard
-- Latency gets yellow when above 100ms and red when above 400ms
+- support for webexec clipboard commands, e.g. use `webexec copy < file.txt` to copy a file to the client's clipboard
+- latency gets yellow when above 100ms and red when above 400ms
+- you can hear the bell
 
 ### Fixed
 
+- peerbook connection on android
 - stuck TWR when chossing not to copy the install command
-- bell is now working
 - latency display is on the gate
 - not asking if "localhost" is the right addresss
 - help text